mic: WIP: some more microphone experimentation
The sample rate seems off and the amplitude seems too low.
Probably not the correct parameters for the DFSDM peripheral, but not
sure how to find them using the ST Excel sheet.

Tested with Interrupt mode, as DMA seems to only support int16_t and
int32_t, but we need 8 bits for the model.

Signed-off-by: Stefan Gloor <code@stefan-gloor.ch>